refactor: modal details directory
