+++ src/app/(provider)/(root)/free-meals/map/_components/StoreDetailModal.tsx
@@ -1,4 +1,4 @@
-import StoreDetails from "@/components/Modal/components/StoreDetails";
+import StoreDetails from "@/app/(provider)/(root)/free-meals/map/_components/StoreDetails";
 import Modal from "@/components/Modal/Modal";
 import { Tables } from "@/supabase/database.types";
 import { useModal } from "@/zustand/modal.store";
