+++ src/app/(provider)/(root)/free-meals/map/_components/StoreDetails.tsx
